failing: test usage of hashes when copying Block
Ran into this issue using Claimchain with a custom FileStore
backend.

We store the content of blocks in files and then load them.
Loading them requires instantiating the block from the values.
Since dict order is non-deterministic and the hid seems to
depend on it this leads to errors.

This is just a failing test for now.