sort dicts before calculating hashes
Dicts have no defined order.
Therefore the hid of a Block should be independent of dict order.
That is two Blocks who are the same except for dict order
should have the same hid.

We achieve this by replacing all dicts with sorted lists
in the Block.items datastructure.
It's the only one that contains dicts.